[refactor] Replace BatchAssetRecordLoader with DataLoader pattern (#24276)
## Summary & Motivation

This replaces BatchAssetRecordLoader with the DataLoader pattern. This result in code getting a fair amount simpler in a lot of places.

Note that this doesn't yet take advantage of any async-io features, it's purely using the caching layer at the moment. However, this sort of refactor would be a necessary step in the path to using asyncio anyway, so this feels like a good step forward.

This also paves the way for AssetGraphView to become a LoadingContext, and take advantage of all this automatic caching stuff, as we'll have several use cases which all depend on a similar set of objects (e.g. RunRecords, AssetRecords, AssetCheckRecords)

## How I Tested These Changes

## Changelog

NOCHANGELOG

- [ ] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_