Allow getting parent/child subsets with AssetCheckKeys (#24359)
## Summary & Motivation

As title. This lets us do partition mappings to / from asset checks, meaning:

1. you can find the upstream partitions of the target asset from that asset's check
2. you can find the downstream subset of a given target check from that check's asset

At the moment, all asset checks are unpartitioned, and so these subsets will always be all or none. However, by modeling it in this way, we don't need to write any custom code to make it work this way, and can just rely on the existing partition mapping logic.

If in the future we decide that asset checks can be partitioned, then no changes will need to be made to these code paths.

## How I Tested These Changes

## Changelog

NOCHANGELOG

- [ ] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_