[refactor] Simplify AutomationConditionEvaluator (#24426)
## Summary & Motivation

Refactors the `.create()` method on the context objects and reorganizes the code a bit more to my liking.

There's an unfortunate complication here caused by a weird thing that we do with non-subsettable multi assets, which is that if any asset decides to materialize, we materialize all of its neighbors as well.

If I could go back in time, I might not have implemented this behavior at all, but I think for now we're essentially stuck with it. Future attempts may be made to hide this nasty logic away more thoroughly.

## How I Tested These Changes

## Changelog

NOCHANGELOG

- [ ] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_