AssetDaemon can produce runs targeting EntitySubset[AssetCheckKey] (#24429)
## Summary & Motivation

As title -- this updates the return type to be Sequence[EntitySubset[EntityKey]], and updates the method for building run requests to process check keys in addition to asset keys.

Does not handle the case when this is running in "daemon mode" (i.e. checks can exist in different code locations), which is an explicit choice to avoid unnecessary complexity. The default behavior is for AutomationConditions to be evaluated against a single code location at a time, and so this is not an issue.

Tests will be done upstack.

## How I Tested These Changes

## Changelog

NOCHANGELOG

- [ ] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_