AutomationConditionEvaluator operate over EntityKeys (#24430)
## Summary & Motivation

This updates the BaseAssetGraph class to produce a dependency graph of all entities (not just AssetKeys), allowing us to topologically sort the entire graph.

Note that I am using `parent_entity_keys` instead of `parent_keys`, because we use `parent_keys` / `child_keys` all over the codebase and historically this has just meant asset keys.

Worth considering renaming those to `parent_asset_keys` and `child_asset_keys`, but that felt like it was outside the scope of this PR.

## How I Tested These Changes

## Changelog

NOCHANGELOG

- [ ] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_