Generic serialized AutomationCondition objects (#24454)
## Summary & Motivation

This allows the serialized objects within the system to be generic (i.e. support either assets or asset checks).

This does not implement a method for storing AutomationConditionEvaluation objects that target AssetCheckKeys, as this will be handled in the next PR in the stack.

## How I Tested These Changes

## Changelog

Insert changelog entry or "NOCHANGELOG" here.

- [ ] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_