Include AssetChecks in AutomationConditionSensorDefinitions (#24465)
## Summary & Motivation

This allows AutomationConditionSensorDefinitions to process AssetChecks by including them in the asset selection.

For now, we will not allow for cases where an asset and its checks are handled by separate sensors. It's unclear to me if we would ever want this, but certainly at the moment our AssetSelection syntax is not really designed for this sort of use case.

## How I Tested These Changes

## Changelog

NOCHANGELOG

- [ ] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_