default the value of run_ids to None for RunsFilter (#24597)
## Summary & Motivation
this change is motivated by some work i'm doing on the runs feed where i
need to replace one of the values in the RunsFilter. However, the
current `__new__` for the runs filter assigns `[]` as the value for
`run_ids` if None is passed. When `copy(filters,
created_before=updated_created_before)` is called, a new `RunsFilter` is
constructed with `[]` passed as the value for `run_ids` which causes the
`check` to fail.

associated internal pr https://github.com/dagster-io/internal/pull/11683


## How I Tested These Changes
existing tests

## Changelog

NOCHANGELOG