Patch kubernetes Api Client to work around lock contention issues
Unclear whether the benefits are worth the risk of breaking changes here, but this is a workaround for the signifcant lock contention issues with the k8s client surfaced in https://github.com/dagster-io/dagster/discussions/23933#discussioncomment-10707901. This appears to be a known issue with the k8s python client - see https://github.com/kubernetes-client/python/issues/2284.

Test Plan: BK