[definition-tags] Add `tags` param to SensorDefinition and subclasses/decorators (#24377)
## Summary & Motivation

Add new `tags` parameter to `SensorDefinition` and associated classes.

The `tags` parameter represents strictly definition tags-- it is under
no circumstances automatically attached to runs emitted from the sensor.
This can be interpreted as matching the behavior on `ScheduleDefinition`
(added upstack), where `tags` are _only_ auto-appended as run tags if a
custom evaluation function is not provided (this is for backcompat
reasons). We don't have to worry about this case with sensors since
there are no instances without custom evaluation functions.

Note that there is one sensor type (`AutomationConditionSensor`) that
already has a `run_tags` parameter. That did not interfere with this PR
since it is under the `run_tags` name, but it's odd that this is the
only sensor type with `run_tags` and warrants discussion elsewhere.

## How I Tested These Changes

Modified existing unit tests.

## Changelog

`SensorDefinition` and all of its variants/decorators now accept a
`tags` parameter. The tags annotate the definition and can be used to
search and filter in the UI.

- [x] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_