[definition-tags] Add definition tags to `ScheduleDefinition` (#24378)
## Summary & Motivation

Tweak the behavior of `tags` on `ScheduleDefinition` to make them act as
"definition tags", as opposed to "run tags".

Downstack we add `tags` to `SensorDefinition`. This was straightforward
because there was no pre-existing `tags` param. The situation is more
complex for schedules-- `ScheduleDefinition` has a preexisting `tags`
parameter that effectively acts as "run tags", i.e. if set it is
attached to runs launched from the schedule.

However, if `tags` _and_ a custom execution function are set, then
`ScheduleDefinition` errors on construction. Further, due to an apparent
past mistake, if `tags` is passed to the `@schedule` decorator, then it
is simply ignored in all but the case where the schedule returns a dict
run config (an old and possibly not-even-supported anymore API)?

This actually works in our favor in this PR, where the aim is for `tags`
to represent definition tags only as often as possible. New behavior:

- We relax the constraint where `ScheduleDefinition` cannot receive both
`tags` and a custom `evaluation_fn`. Instead, if both of these are set,
the `tags` act as pure definition tags-- they are not attached to any
emitted `RunRequest`.
- If no custom evaluation fn is provided, then tags both act as
definition tags _and_ get automatically appended to emitted
`RunRequests`. This isn't great but is necessary for backcompat.

NOTE: though a `tags` param was already present on `ScheduleDefinition`,
it wasn't exposed on the remote representation. This PR adds that
exposure, which is necessary for the tags to function as "definition
tags".

## How I Tested These Changes

New unit tests.

## Changelog

It is now possible to set both `tags` and a custom `execution_fn` on a
`ScheduleDefinition`. Schedule `tags` are intended to annotate the
definition and can be used to search and filter in the UI. They will not
be attached to run requests emitted from the schedule if a custom
`execution_fn` is provided. If no custom `execution_fn` is provided,
then for back-compatibility the tags will also be automatically attached
to run requests emitted from the schedule.

- [x] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_