Expose a global 'max_runtime_seconds' field on run monitoring (#24653)
## Summary & Motivation
Allows you to set a global max across all runs in the deployment instead
of tagging individual runs

## How I Tested These Changes
BK

## Changelog
- [x] `NEW` Added a `max_runtime_seconds` configuration option to run
monitoring, allowing you to specify that any run in your Dagster
deployment should terminate if it exceeds a certain runtime. Prevoiusly,
jobs had to be individually tagged with a `dagster/max_runtime` tag in
order to take advantage of this feature. Jobs and runs can still be
tagged in order to override this value for an individual run.