add AssetSpec.with_io_manager_key and update SourceAsset deprecation warning (#24690)
## Motivation

`SourceAsset` has been deprecated in favor of `AssetSpec`. Unlike
`SourceAsset`, `AssetSpec` has no `io_manager_key` attribute, but an IO
manager key can be provided via metadata to dictate how the asset is
loaded in downstream assets.

This metadata-based IO management has up until this point not been very
discoverable (you need to read the migration guide or
https://docs.dagster.io/concepts/io-management/io-managers#using-io-managers-to-load-source-data
to find it).

## How I Tested These Changes

## Changelog

`AssetSpec` now has a `with_io_manager_key` method that returns an
`AssetSpec` with the appropriate metadata entry to dictate the key for
the IO manager used to load it. The deprecation warning for
`SourceAsset` now references this method.

- [x] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_