Add default_condition to AutomationConditionSensorDefinition (#24613)
## Summary & Motivation

Adds a (hidden) default_condition argument to AutomationConditionSensorDefinition.

Currently, this only works for sensors that evaluated in user-space, as otherwise we'd need to serialize the AutomationCondition somewhere in the SensorSnap. Given that we want to move to having all of these sensors operate in user-space, it seems like the better call is to avoid bloating up the SensorSnaps for a feature that we don't want to support indefinitely.

## How I Tested These Changes

## Changelog

NOCHANGELOG

- [ ] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_