Refactor airlift defintions construction logic for more elegant replacement (#24665)
## Summary & Motivation

I had encountered a bug around using serdes objects in the DefinitionsLoadContext. I ended up doing this refactor and adding a test case to understand the existing behavior. This made the replacement upstack more minimal so submitting a separate PR.

Specifically `defs_with_airflow_metadata_applied` was kind of in an unexpected spot (deep in the callstack) so moved it much closer to where the state is actually fetched.

## How I Tested These Changes

BK

## Changelog

NOCHANGELOG