Add functional type mapping object to control complexity in compute_serialized_data (#24680)
## Summary & Motivation

One of the big sources of complexity in airlift right now is `compute_serialized_data` which iteratively builds up a bunch of data structures in parallel. If we want to change anything about the way this works, it is complicated and error-prone.

Instead of building up these data structures imperatively in one pass, we instead will construct data structures that use functional programming to build up "views" of the data that are easier to reason about. 

In this case we introduce `TaskSpecMappingInfo` which manages all the various data that maps task_id and dag_ids to asset keys. 

It also accepts a list of asset specs, so the operations are guaranteed to be fast, and all results are memoized with `cached_property`.

## How I Tested These Changes

BK

## Changelog

NOCHANGELOG