FetchedAirflowData (#24681)
## Summary & Motivation

Similar in spirit to the upstack #24680. This adds another functional "view" object, but this time over prefetched data from the Airflow instance.

This current PR changes no behavior w/r/t to interaction with the AirflowInstance, but it structures things in way where it will be trivial to change fetching task infos to bulk fetching.

## How I Tested These Changes

BK

## Changelog

NOCHANGELOG