add "dagster-dbt/materialization_type" to default metadata produced for dbt assets (#23552)
## Summary & Motivation

Without this, it's difficult to determine whether a dbt model is a view
or not.

https://docs.getdbt.com/docs/build/materializations

One question that came up is whether this should instead be
"dagster-dbt/**materialized**_type", given that it's configured in dbt
using e.g. `+materialized: table`. However, in their docs they call it
"materialization", and the past tense seems a little awkward for this
use case.

## How I Tested These Changes