remove SerializedAssetKeyScopedAssetGraphData and SerializedAssetKeyScopedData (#24729)
## Summary & Motivation

With #24728 we no longer compute check keys associated with the asset keys in the serialization process and instead compute them on demand in the sensor.

It was actually the only piece of data we accessed in the `SerializedAssetKeyScopedAssetGraphData` which allowed for a set of cascading deletes.

Also was able to eliminate the `GenericNonScalarKeyMappingSerializer` and instead just rely on serializing key-value records (I tried tuples at first which would be even less code. Not sure what is going on there.). We can very easily convert lists to dictionaries and back, which is simpler and more obvious.

e.g.

```python
@record
@whitelist_for_serdes
@record
class SerializedAirflowDefinitionsData:
    key_scoped_data_items: List[KeyScopedDataItem]

    @cached_property
    def key_scope_data_map(self) -> Mapping[AssetKey, "SerializedAssetKeyScopedAirflowData"]:
        return {item.asset_key: item.data for item in self.key_scoped_data_items}
```

and

```python
    return SerializedAirflowDefinitionsData(
        key_scoped_data_items=[
            KeyScopedDataItem(asset_key=k, data=v)
            for k, v in fetched_airflow_data.airflow_data_by_key.items()
        ],
    )
```



## How I Tested These Changes

## Changelog

Insert changelog entry or "NOCHANGELOG" here.

- [ ] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_