Expose AssetSpec on OutputContext (#24495)
## Summary & Motivation

Contributes to https://github.com/dagster-io/dagster/issues/23965. 

Exposes the `AssetSpec` on the `OutputContext`, which in turn contains
things like the `code_version` and `partitions_definition`.

Some thoughts:
- The asset `PartitionsDefinition` and `AssetKey` are already exposed on
the `OutputContext` directly, but they are also part of the `AssetSpec`.
Maybe this makes the directly exposed properties superfluous in the long
run?
- The way to get to the `AssetSpec` via
`step_context.job_def.asset_layer` seems like a bit of a long way
around. Perhaps the `AssetSpec` should be part of the `OutputContext`
constructor and be attached directly as a property?

## How I Tested These Changes

Added a unit test to confirm that the `AssetSpec` is exposed on the
`OutputContext` by checking with an io manager

## Changelog

`OutputContext` now exposes the `AssetSpec` of the asset that is being
stored as an output

- [x] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_