[backcompat-tests] Add list_repo test (#24523)
## Summary & Motivation

- Add an asset, schedule, and sensor to the backcompat test repo
- Add a test that queries the contents of the repo
- Bump oldest tested Dagster version to 0.14.0 (Feb 17, 2022) from
0.12.8 (Aug 27, 2021). I did this because 0.12.8 was choking on the
asset, schedule, sensor I added. Couldn't discern the exact reason but I
believe it was because `@asset` did not yet exist. An alternative
solution would be to add the defs in a new repo that is only used our
"latest release" tests.

The motivation is to generally strengthen our backcompat tests and
support upstack PRs that are doing heavy renaming in the external/snap
layer.

## How I Tested These Changes

BK

## Changelog

NOCHANGELOG

- [ ] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_