add `exclude_subruns` option to GrapheneRunsFilter (#24634)
## Summary & Motivation
Adds `excludeSubruns` to `GrapheneRunsFilter` so that the UI can power
the "Show runs within backfills" checkbox

An alternate implementation would be to have a separate parameter for
the `runsFeedOrError` query that determines whether we exclude subruns.
I tried that [here](https://github.com/dagster-io/dagster/pull/24710)
but ran into some confusing gql errors. Marco had a slight preference
for implementing the UI as passing the option to the RunsFilter, so i
switched to that approach. I can go back to the parameter version though
if we want to be more cautious about adding this filter to more APIs

Corresponding internal PR
https://github.com/dagster-io/internal/pull/11793

## How I Tested These Changes

## Changelog

NOCHANGELOG