[ui] Fix Automations page when only schedules present (#24887)
## Summary & Motivation

When determining whether to show a code location bucket in the Automations page, repos with 1+ schedules and zero sensors were being incorrectly ignored. The result was an empty Automations page even though there were schedules that should have appeared.

## How I Tested These Changes

View an org with a single code location that has schedules but not sensors. Verify that the Automations page now shows them correctly.

## Changelog

- [ ] `BUGFIX` In new navigation experimental flag, fix automations page for code locations that have schedules but no sensors.