support backfill id filtering for backfills (#24799)
## Summary & Motivation
Adds a `backfill_ids` filter to `BulkActionsFilter` so that the
`backfill id` filter option on the runs feed can operate on backfills
too.

I originally considered just adding a special case when fetching the
runs feed where if the backfill id filter was passed we directly fetched
the backfill. however, that required manually checking that all of the
other filters matched the backfill, and that the cursor was respected,
which felt prone to bugs. You can see the draft here if interested
https://github.com/dagster-io/dagster/pull/24793

corresponding internal pr
https://github.com/dagster-io/internal/pull/11818

## How I Tested These Changes

## Changelog

NOCHANGELOG