Terminate execute-step process early when it runs on a run that is canceling or otherwise finisehed (#24896)
## Summary & Motivation
This should be rare, but can happen in, for example, the sequence of
events in https://github.com/dagster-io/dagster/issues/24889. If a step
process manages to start while the run is in a CANCELING or CANCELED or
FAILURE state, short circuit it early.

## How I Tested These Changes
BK (new test case)

## Changelog

- [x] `BUGFIX` [dagster-k8s] Fixed an issue where run termination
sometimes did not terminate all step processes when using the
`k8s_job_executor`, if the termination was initiated while it was in the
middle of launching a step pod.