Allow DA to selectively emit backfills (#24905)
## Summary & Motivation

The previous behavior when enabling DA to emit backfills was to *always* emit a backfill regardless of it was required or not.

This updates the behavior to only emit a backfill when more than one partition is requested for a given asset at one time.

If a backfill is required, the system will group together all "touching" assets into a single backfill to ensure that execution order is respected.

This means that it is possible for the system to request multiple backfills and multiple runs on the same tick.

## How I Tested These Change

## Changelog

NOCHANGELOG

- [ ] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_