[dagster-airlift] move defs/mapping construction out of AirflowDefinitionsData (#24897)
## Summary & Motivation
Part of the stack to make AirflowDefinitionsData user-invokable. This
moves the mapping functions to operate purely on the serialized data. I
opted not to make these methods properties on the serialized data so not
as to pollute the "serialized_data" file, which I think is very coherent
right now.

## How I Tested These Changes
Existing tests

## Changelog
NOCHANGELOG