[dagster-airlift] change AirflowDefinitionsData to operate on fully resolved definitions (#24900)
## Summary & Motivation
Changes parameterization to retrieve serialized data from Definitions. 

Requires an additional pr for the instance name. 
A downstream will change this to only look at the asset metadata,
instead of relying on the serialized metadata (except for migration
state for dags/tasks, which we're waiting on the airflow conf to change)

## How I Tested These Changes
Existing tests

## Changelog
NOCHANGELOG