[ui] Update nav flag behavior (#24976)
## Summary & Motivation

A few changes to the behavior of the new IA flag.

- Restore the left nav, and only show it by sliding over the main
content, even for large screens.
- Allow immediately turning off the flag via the Alert at the top of the
app.
- Restore the gear button in the top nav for users in the flag.

## How I Tested These Changes

Load the app with the flag on and off, verify that the left nav and gear
button render and behave as expected. Click to turn the flag off via the
Alert, verify that it works correctly.

## Changelog

NOCHANGELOG