stop ignoring AssetsDefinitions with only checks passed to Definitions (#24979)
## Summary & Motivation

In the course of implementing `map_asset_specs`, I noticed an oddity /
gap in how we handle canonicalized asset check definitions.

During construction of `Definitions`, we canonicalize
`AssetChecksDefinition` objects to `AssetsDefinition` objects. However,
if these are then passed in to construct a new `Definitions` object,
then they are ignored. This PR addresses that.

## How I Tested These Changes

## Changelog

Insert changelog entry or "NOCHANGELOG" here.

- [ ] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_