Moving synthetic materialization generation to not be passed on proxied state (#24981)
## Summary & Motivation

This PR changes that way that we determine whether or not to generate asset materializations for completed tasks.

Previously were we relying on state fetched at code location load time to make this determination. Unfortunately, this meant that as that state changed and the code location was not reloaded, things would get out of sync and we would create phantom materializations.

This relies on a new scheme that uses dagster runs as the source of truth.

When the sensor consumes an airflow dag run, it finds all the task instances that have successfully completed.  It also queries the dagster metadata store for all dagster runs tagged with that airflow dag run's ID, which was attached by the proxied operator.

If, for a given completed, successful task_id, there exists a Dagster run, we know that there has been a best effort to materialize the associated asset via a proxy operation. If the dagster run does not, then the execution happened within Airflow, and we should synthesize a materialization and attach airflow metadata information to it.

## How I Tested These Changes

Manual testing with "kitchen-sink" and automated testing.

Testing I would say is relatively weak here as we don't do this at massive scale and for a wide swath of use cases where there is weird state in the Airflow instance.

## Changelog

NOCHANGELOG