Add execute_k8s_job arg to delete failed k8s jobs
The current behavior of execute_k8s_job is to immediately delete a
failed Kubernetes job. This change adds a boolean parameter to control
this behavior so that users and supporting processes (such as log
ingestion tools) have time to access a failed job's pods. The new arg
defaults to True, preserving existing behavior.