Update first tick behavior of AutomationCondition.eager() (#24700)
## Summary & Motivation

Updates the behavior of the `newly_true()` condition to optionally change its "baseline" assumption of what was true before the first evaluation. This allows us to better fulfill the promise of "only reacting to changes which happened after the condition was added to the asset", as we do for the "deps updated" part of the condition.

In short, this provides a strong bit of protection against the "surprise backfill" issue, without having to add any new rate-limiting concepts to the system.

## How I Tested These Changes

## Changelog

AutomationCondition.eager() will now only launch runs for missing partitions which become missing *after* the condition has been added to the asset. This avoids situations in which the eager policy kicks off a large amount of work when added to an asset with many missing historical static/dynamic partitions.

- [x] `NEW` _(added new feature or capability)_
- [ ] `BUGFIX` _(fixed a bug)_
- [ ] `DOCS` _(added or updated documentation)_