[record] add replace method with namedtuple._replace behavior (#24894)
Currently, `LegacyNamedTupleMixin._replace` is setup using `copy` which
goes through the `@record` constructor, using the manually implemented
`__new__` if present.

This actually diverges from `namedtuple._replace` which bypasses any
custom `__new__` using the tuple `__new__` directly via `_make`.

We had an issue where a callsite using `_replace` was believed to be a
`namedtuple` but was actually a `@record` wtih `LegacyNamedTupleMixin`
and this behavior divergence caused an issue. Validation that was
expected to be skipped was run and caused an exception.

Avoid this by creating `replace` an alternative to `copy` which bypasses
any custom `__new__`. This can be useful when trying to skip certain
validation or if the custom `__new__` diverges from the record field set
substantially and its hard to get back to the original type that the
custom `__new__` takes. For example triyng to copy a `RepositoryHandle`
withtout having the `CodeLocation` it was originally derived from.

## How I Tested These Changes

added tests