feat: graphql add Run.hasRunMetricsEnabled field (#24662)
## Summary & Motivation

This change introduces a new field `hasRunMetricsEnabled` to the GrapheneRun type in the GraphQL schema. It allows clients to query whether run metrics are enabled for a specific run. The implementation checks for the presence of run metric tags and their boolean values.

This enables hiding the run metrics tag to declutter the frontend of this implementation detail.

## How I Tested These Changes

- BK w/ added test

## Changelog

NOCHANGELOG