[definition-tags] Eliminate the NormalizedTags construct (#24789)
## Summary & Motivation

Remove the `NormalizedTags` construct from the code. I understand the
theory behind this construct (that tags need only be normalized once)
but it was used inconsistently in the code and isn't a pattern we use
elsewhere. The net effect was to add noise to anything one wants to do
with tags.

## How I Tested These Changes

Existing test suite.

## Changelog

NOCHANGELOG