[dagster-airlift] Add python 3.8 support (#25151)
## Summary & Motivation
In order to support 3.8, we need to place an upper bound on boto3 (older versions don't have access to mwaa).
The version I chose is quite old, I don't anticipate this causing a problem. I guess with some resolvers it chooses a version pre-mwaa though, as that's why the tests were failing.
## How I Tested These Changes
Tests now run on 3.8
## Changelog
NOCHANGELOG