[AssetPageHeader] Add keys to breadcrumb elements to fix crazy flashing (#25188)
## Summary & Motivation

I'm honestly not entirely sure what is happening here but while
debugging I noticed that the internal React fiber nodes were being
improperly reused for different instances of the breadcrumb. In one
render the instance would be rendering "SEO", and in the next it was
rendering "Mart". I'm assuming this is due to our recent Blueprint
upgrade. Adding keys onto the breadcrumb elements fixes the issue.


https://app.usepylon.com/issues?conversationID=6713c615-56be-43b1-9559-18d76f2f5c32

## How I Tested These Changes

Successfully Loaded an asset page that was previously flickering like
crazy.


## Changelog

[Bugfix][ui] Fixed an issue where the breadcumb on asset pages would
flicker nonstop.