[dagster-embedded-elt] fix extraction of metadata for dlt assets whose source and destination identifiers differ (#25266)
## Summary & Motivation
This fixes a `KeyError` when using `resource.table_name` to extract
metadata. The
[docs](https://dlthub.com/docs/general-usage/naming-convention#source-identifiers-vs-destination-identifiers)
make the distinction between source and destination identifiers. For
example, if `DltResource.table_name` is defined to be `some__table`,
then the final table name after the normalization step will actually be
`some_table`, which leads to the `KeyError`.

- Instead of using `resource.table_name` directly, we can normalize it
using the same naming convention used by the dlt pipeline. This part was
taken from the
[`DltSource.discover_schema()`](https://github.com/dlt-hub/dlt/blob/8798c1753a529db1d78d3fd4189e552c3705d250/dlt/extract/source.py#L306)
method.
- I also added a `try except` block just to avoid similar issues in the
future since the metadata is not necessary for the asset to materialize.

Note that this also fixes the extraction of other metadata that were
previously missing if the source and destination identifiers differed.
However, it doesn't raise an error in that case.

## How I Tested These Changes
Added `table_name=RepoIssues` to test that we are able to extract
metadata for table names that will be transformed during the
normalization step.

## Changelog

[dagster-embedded-elt] fix extraction of metadata for dlt assets whose
source and destination identifiers differ